+++ node/clients/checkout.ts
@@ -1,14 +1,11 @@
-import { InstanceOptions, IOContext, JanusClient, RequestConfig } from '@vtex/api'
+import {
+  InstanceOptions,
+  IOContext,
+  JanusClient,
+  RequestConfig,
+  IOResponse,
+} from '@vtex/api'
 import { checkoutCookieFormat, statusToError } from '../utils'
-
-export interface SimulationData {
-  country: string
-  items: { id: string; quantity: number | string; seller: string }[]
-  postalCode?: string
-  isCheckedIn?: boolean
-  priceTables?: string[]
-  marketingData?: Record<string, string>
-}
 
 export class Checkout extends JanusClient {
   public constructor(ctx: IOContext, options?: InstanceOptions) {
@@ -29,7 +26,7 @@ export class Checkout extends JanusClient {
     return {
       Cookie: `${checkoutCookie}vtex_segment=${
         this.context.segmentToken
-      };vtex_session=${this.context.sessionToken};`,
+        };vtex_session=${this.context.sessionToken};`,
     }
   }
 
@@ -41,7 +38,7 @@ export class Checkout extends JanusClient {
   }
 
   public addItem = (orderFormId: string, items: any) =>
-    this.post(
+    this.post<OrderForm>(
       this.routes.addItem(orderFormId, this.getChannelQueryString()),
       { orderItems: items },
       { metric: 'checkout-addItem' }
@@ -116,25 +113,24 @@ export class Checkout extends JanusClient {
 
   public addAssemblyOptions = async (
     orderFormId: string,
-    itemId: string,
+    itemId: string | number,
     assemblyOptionsId: string,
     body: any
   ) =>
-    this.post(
+    this.post<OrderForm>(
       this.routes.assemblyOptions(orderFormId, itemId, assemblyOptionsId),
       body,
       { metric: 'checkout-addAssemblyOptions' }
     )
-
   public removeAssemblyOptions = async (
     orderFormId: string,
-    itemId: string,
+    itemId: string | number,
     assemblyOptionsId: string,
     body: any
   ) =>
-    this.delete(
+    this.delete<OrderForm>(
       this.routes.assemblyOptions(orderFormId, itemId, assemblyOptionsId),
-      { metric: 'checkout-removeAssemblyOptions', params: body }
+      { metric: 'checkout-removeAssemblyOptions', data: body }
     )
 
   public updateOrderFormCheckin = (orderFormId: string, checkinPayload: any) =>
@@ -142,20 +138,40 @@ export class Checkout extends JanusClient {
       metric: 'checkout-updateOrderFormCheckin',
     })
 
-  public orderForm = (useRaw?: boolean) => {
-    const method = useRaw ? this.postRaw : this.post
-    return method(
+  public orderForm = () => {
+    return this.post<OrderForm>(
       this.routes.orderForm,
       { expectedOrderFormSections: ['items'] },
       { metric: 'checkout-orderForm' }
     )
   }
 
+  public orderFormRaw = () => {
+    return this.postRaw<OrderForm>(
+      this.routes.orderForm,
+      { expectedOrderFormSections: ['items'] },
+      { metric: 'checkout-orderForm' }
+    )
+  }
+
+  public changeToAnonymousUser = () => {
+    const { orderFormId } = this.context as CustomIOContext
+
+    if (!orderFormId) {
+      throw new Error('Missing orderFormId. Use withOrderFormId directive.')
+    }
+
+    return this.get(
+      this.routes.changeToAnonymousUser(orderFormId),
+      { metric: 'checkout-change-to-anonymous' }
+    )
+  }
+
   public orders = () =>
     this.get(this.routes.orders, { metric: 'checkout-orders' })
 
-  public simulation = (simulation: SimulationData) =>
-    this.post(
+  public simulation = (simulation: SimulationPayload) =>
+    this.post<OrderForm>(
       this.routes.simulation(this.getChannelQueryString()),
       simulation,
       {
@@ -168,18 +184,20 @@ export class Checkout extends JanusClient {
       ...config.headers,
       ...this.getCommonHeaders(),
     }
-    return this.http.get<T>(url, config).catch(statusToError)
+    return this.http.get<T>(url, config).catch(statusToError) as Promise<T>
   }
 
-  protected post = (url: string, data?: any, config: RequestConfig = {}) => {
+  protected post = <T>(url: string, data?: any, config: RequestConfig = {}) => {
     config.headers = {
       ...config.headers,
       ...this.getCommonHeaders(),
     }
-    return this.http.post<any>(url, data, config).catch(statusToError)
+    return this.http.post<T>(url, data, config).catch(statusToError) as Promise<
+      T
+    >
   }
 
-  protected postRaw = async (
+  protected postRaw = async <T>(
     url: string,
     data?: any,
     config: RequestConfig = {}
@@ -188,7 +206,9 @@ export class Checkout extends JanusClient {
       ...config.headers,
       ...this.getCommonHeaders(),
     }
-    return this.http.postRaw<any>(url, data, config).catch(statusToError)
+    return this.http
+      .postRaw<T>(url, data, config)
+      .catch(statusToError) as Promise<IOResponse<T>>
   }
 
   protected delete = <T>(url: string, config: RequestConfig = {}) => {
@@ -196,7 +216,9 @@ export class Checkout extends JanusClient {
       ...config.headers,
       ...this.getCommonHeaders(),
     }
-    return this.http.delete<T>(url, config).catch(statusToError)
+    return this.http.delete<T>(url, config).catch(statusToError) as Promise<
+      IOResponse<T>
+    >
   }
 
   protected patch = <T>(
@@ -208,7 +230,9 @@ export class Checkout extends JanusClient {
       ...config.headers,
       ...this.getCommonHeaders(),
     }
-    return this.http.patch<T>(url, data, config).catch(statusToError)
+    return this.http
+      .patch<T>(url, data, config)
+      .catch(statusToError) as Promise<T>
   }
 
   protected put = <T>(url: string, data?: any, config: RequestConfig = {}) => {
@@ -216,7 +240,9 @@ export class Checkout extends JanusClient {
       ...config.headers,
       ...this.getCommonHeaders(),
     }
-    return this.http.put<T>(url, data, config).catch(statusToError)
+    return this.http.put<T>(url, data, config).catch(statusToError) as Promise<
+      T
+    >
   }
 
   private get routes() {
@@ -239,7 +265,7 @@ export class Checkout extends JanusClient {
         `${base}/orderForm/${orderFormId}/attachments/${field}`,
       assemblyOptions: (
         orderFormId: string,
-        itemId: string,
+        itemId: string | number,
         assemblyOptionsId: string
       ) =>
         `${base}/orderForm/${orderFormId}/items/${itemId}/assemblyOptions/${assemblyOptionsId}`,
@@ -249,6 +275,8 @@ export class Checkout extends JanusClient {
       orders: `${base}/orders`,
       simulation: (queryString: string) =>
         `${base}/orderForms/simulation${queryString}`,
+      changeToAnonymousUser: (orderFormId: string) =>
+        `/checkout/changeToAnonymousUser/${orderFormId}`
     }
   }
 }