+++ node/resolvers/checkout/attachmentsHelper.ts
@@ -1,4 +1,17 @@
-import { all, filter, find, partition, path, pathOr, propEq } from 'ramda'
+import {
+  all,
+  filter,
+  find,
+  partition,
+  path,
+  pathOr,
+  propEq,
+  omit,
+  eqProps,
+  map,
+  compose,
+  equals,
+} from 'ramda'
 
 export const CHOICE_TYPES = {
   MULTIPLE: 'MULTIPLE',
@@ -6,95 +19,137 @@ export const CHOICE_TYPES = {
   TOGGLE: 'TOGGLE',
 }
 
-interface OptionsType {
-  id: string
-  quantity: number
-  assemblyId: string
-  seller: string
+const getNewItemsOnly = (
+  previousItems: OrderFormItem[],
+  allItems: OrderFormItem[]
+) => {
+  return allItems.filter(item => !previousItems.find(eqProps('uniqueId', item)))
 }
 
-interface OptionRequestParam {
-  id: string
-  seller: string
-}
+const findRecentlyAddedParent = (
+  recentlyAdded: OrderFormItem[],
+  id: string,
+  assemblyId: string | null
+) =>
+  recentlyAdded.find(i => i.id === id && i.parentAssemblyBinding === assemblyId)
 
-interface OptionRequestAddParam extends OptionRequestParam {
-  quantity: number
-}
-
-interface ItemsToAdd {
-  id: string
-  quantity: number
-  seller: string
-  index?: number
-  options?: OptionsType[]
-}
+type OptionRequestUnit = Omit<AssemblyOptionInput, 'assemblyId'>
 
 interface AddOptionsLogicInput {
   checkout: Context['clients']['checkout']
-  orderFormId: string
-  itemIndex: string
-  options?: OptionsType[]
+  orderForm: OrderForm
+  itemIndex: string | number
+  options?: AssemblyOptionInput[]
+  oldItems: OrderFormItem[]
 }
 
-const addAssemblyBody = (items: OptionRequestAddParam[]) => ({
+const addAssemblyBody = (items: OptionRequestUnit[]) => ({
   composition: {
-    items,
+    items: map<OptionRequestUnit>(omit(['options']) as any, items),
   },
   noSplitItem: true,
 })
 
-const removeAssemblyBody = (items: OptionRequestAddParam[]) => ({
+const removeAssemblyBody = (items: OptionRequestUnit[]) => ({
   composition: {
-    items: removeQuantity(items),
+    items: map<OptionRequestUnit>(omit(['quantity', 'options']) as any, items),
   },
 })
 
-const removeQuantity = (
-  options: OptionRequestAddParam[]
-): OptionRequestParam[] => options.map(({ seller, id }) => ({ seller, id }))
+const joinOptionsWithType = (options: AssemblyOptionInput[]) => {
+  const result = {} as Record<string, OptionRequestUnit[]>
+  for (const option of options) {
+    const { assemblyId, ...rest } = option
+    const currentArray = result[assemblyId] || []
+    currentArray.push(rest)
+    result[assemblyId] = currentArray
+  }
+  return result
+}
 
-const joinOptionsWithType = (options: OptionsType[]) => {
-  return options.reduce<{ [key: string]: OptionRequestAddParam[] }>(
-    (prev, curr) => {
-      const { assemblyId, ...rest } = curr
-      return {
-        ...prev,
-        [assemblyId]: [...(prev[assemblyId] || []), rest],
-      }
-    },
-    {}
-  )
+const addOptionsRecursive = async (
+  items: OptionRequestUnit[],
+  assemblyId: string,
+  orderForm: OrderForm,
+  oldItems: OrderFormItem[],
+  checkout: Context['clients']['checkout']
+) => {
+  const recentlyAdded = getNewItemsOnly(oldItems, orderForm.items)
+  for (const item of items) {
+    const parentItem = findRecentlyAddedParent(
+      recentlyAdded,
+      item.id!.toString(),
+      assemblyId
+    )
+    const parentIndex =
+      parentItem &&
+      orderForm.items.findIndex(propEq('uniqueId', parentItem.uniqueId))
+    if (parentIndex == null || parentIndex < 0) {
+      continue
+    }
+    await addOptionsLogic({
+      checkout,
+      itemIndex: parentIndex,
+      options: item.options,
+      orderForm,
+      oldItems,
+    })
+  }
 }
 
 const addOptionsLogic = async (input: AddOptionsLogicInput) => {
-  const { checkout, orderFormId, itemIndex, options } = input
+  const { checkout, orderForm, itemIndex, options, oldItems } = input
   if (!options || options.length === 0) {
     return
   }
-  const isRemove = (option: OptionsType) => option.quantity === 0
-  const [toRemove, toAdd] = partition<OptionsType>(isRemove, options)
+  const isRemove = (option: AssemblyOptionInput) => option.quantity === 0
+  const [toRemove, toAdd] = partition<AssemblyOptionInput>(isRemove, options)
   const joinedToAdd = joinOptionsWithType(toAdd)
   const joinedToRemove = joinOptionsWithType(toRemove)
-  for (const [assemblyId, parsedOptions] of Object.entries(joinedToAdd)) {
-    await checkout
-      .addAssemblyOptions(
-        orderFormId,
-        itemIndex,
-        assemblyId,
-        addAssemblyBody(parsedOptions)
-      )
-      .catch(() => null)
-  }
-  for (const [assemblyId, parsedOptions] of Object.entries(joinedToRemove)) {
-    await checkout
+  const idsToAdd = Object.keys(joinedToAdd)
+  const idsToRemove = Object.keys(joinedToRemove)
+  let recentOrderForm = orderForm
+
+  for (const assemblyId of idsToRemove) {
+    const parsedOptions = joinedToRemove[assemblyId]
+    const response = await checkout
       .removeAssemblyOptions(
-        orderFormId,
+        orderForm.orderFormId,
         itemIndex,
         assemblyId,
         removeAssemblyBody(parsedOptions)
       )
-      .catch(() => null)
+      .catch(() => ({ data: recentOrderForm }))
+    recentOrderForm = response.data
+  }
+
+  for (const assemblyId of idsToAdd) {
+    const parsedOptions = joinedToAdd[assemblyId]
+    recentOrderForm = await checkout
+      .addAssemblyOptions(
+        orderForm.orderFormId,
+        itemIndex,
+        assemblyId,
+        addAssemblyBody(parsedOptions)
+      )
+      .catch(() => recentOrderForm)
+  }
+
+  for (const assemblyId of idsToAdd) {
+    const parsedOptions = joinedToAdd[assemblyId]
+    const itemsWithRecursiveOptions = parsedOptions.filter(
+      ({ options }) => !!options
+    )
+
+    if (itemsWithRecursiveOptions.length > 0) {
+      await addOptionsRecursive(
+        itemsWithRecursiveOptions,
+        assemblyId,
+        recentOrderForm,
+        oldItems,
+        checkout
+      )
+    }
   }
 }
 
@@ -107,25 +162,36 @@ const addOptionsLogic = async (input: AddOptionsLogicInput) => {
  */
 
 export const addOptionsForItems = async (
-  items: ItemsToAdd[],
+  items: OrderFormItemInput[],
   checkout: Context['clients']['checkout'],
-  orderForm: any
+  orderForm: OrderForm,
+  oldItems: OrderFormItem[]
 ) => {
+  const recentlyAdded =
+    items.length > 0 ? getNewItemsOnly(oldItems, orderForm.items) : []
   for (const item of items) {
     if (!item.options || item.options.length === 0) {
       continue
     }
-    const parentIndex = orderForm.items.findIndex(
-      (cartItem: any) => cartItem.id.toString() === item.id.toString()
+
+    const parentItem = findRecentlyAddedParent(
+      recentlyAdded,
+      item.id!.toString(),
+      null
     )
-    if (parentIndex < 0) {
+
+    const parentIndex =
+      parentItem &&
+      orderForm.items.findIndex(propEq('uniqueId', parentItem.uniqueId))
+    if (parentIndex == null || parentIndex < 0) {
       continue
     }
     await addOptionsLogic({
       checkout,
       itemIndex: parentIndex,
       options: item.options,
-      orderFormId: orderForm.orderFormId,
+      orderForm,
+      oldItems,
     })
   }
 }
@@ -133,12 +199,12 @@ export const addOptionsForItems = async (
 const filterCompositionNull = (assemblyOptions: AssemblyOption[]) =>
   assemblyOptions.filter(({ composition }) => !!composition)
 
-export const buildAssemblyOptionsMap = (orderForm: any) => {
-  const metadataItems = pathOr(
+export const buildAssemblyOptionsMap = (orderForm: OrderForm) => {
+  const metadataItems = pathOr<[], MetadataItem[]>(
     [],
     ['itemMetadata', 'items'],
     orderForm
-  ) as MetadataItem[]
+  )
   return metadataItems
     .filter(
       ({ assemblyOptions }) => assemblyOptions && assemblyOptions.length > 0
@@ -168,7 +234,7 @@ const isParentOptionToggleChoice = ({ composition }: AssemblyOption) => {
   return all(propEq('maxQuantity', 1))(items)
 }
 
-const getItemChoiceType = (childAssemblyData?: AssemblyOption) => {
+export const getItemChoiceType = (childAssemblyData?: AssemblyOption) => {
   if (!childAssemblyData) {
     return CHOICE_TYPES.MULTIPLE
   }
@@ -185,7 +251,7 @@ const getItemChoiceType = (childAssemblyData?: AssemblyOption) => {
 }
 
 const getItemComposition = (
-  childItem: OrderFormItem,
+  id: string,
   childAssemblyData?: AssemblyOption
 ): CompositionItem | undefined => {
   if (!childAssemblyData) {
@@ -193,7 +259,7 @@ const getItemComposition = (
   }
   const items =
     (childAssemblyData.composition && childAssemblyData.composition.items) || []
-  return find<CompositionItem>(propEq('id', childItem.id), items)
+  return find<CompositionItem>(propEq('id', id), items)
 }
 
 const isSonOfItem = (parentIndex: number) =>
@@ -204,11 +270,17 @@ export const isParentItem = ({
   parentAssemblyBinding,
 }: OrderFormItem) => parentItemIndex == null && parentAssemblyBinding == null
 
+export const getPositionInOrderForm = (
+  { items }: OrderForm,
+  { uniqueId }: OrderFormItem
+) => items.findIndex(orderItem => orderItem.uniqueId === uniqueId)
+
 export const buildAddedOptionsForItem = (
   item: OrderFormItem,
   index: number,
   childs: OrderFormItem[],
-  assemblyOptionsMap: Record<string, AssemblyOption[]>
+  assemblyOptionsMap: Record<string, AssemblyOption[]>,
+  orderForm: OrderForm
 ) => {
   const children = filter<OrderFormItem>(isSonOfItem(index), childs)
   return children.map(childItem => {
@@ -217,7 +289,7 @@ export const buildAddedOptionsForItem = (
       propEq('id', childItem.parentAssemblyBinding)
     )(parentAssemblyOptions)
     const compositionItem = getItemComposition(
-      childItem,
+      childItem.id,
       childAssemblyData
     ) || { initialQuantity: 0 }
     return {
@@ -225,7 +297,15 @@ export const buildAddedOptionsForItem = (
       compositionItem,
       extraQuantity:
         childItem.quantity / item.quantity - compositionItem.initialQuantity,
-      item: childItem,
+      item: {
+        ...childItem,
+        assemblyOptionsData: {
+          index: getPositionInOrderForm(orderForm, childItem),
+          assemblyOptionsMap,
+          childs,
+          orderForm,
+        },
+      },
       normalizedQuantity: childItem.quantity / item.quantity,
     }
   })
@@ -239,7 +319,7 @@ const findInitialItemOnCart = (initialItem: InitialItem) => (
 
 const isInitialItemMissing = (
   parentCartItem: OrderFormItem,
-  orderForm: any
+  orderForm: OrderForm
 ) => (initialItem: InitialItem): RemovedItem | null => {
   const orderFormItem = find(
     findInitialItemOnCart(initialItem),
@@ -273,17 +353,25 @@ interface InitialItem extends CompositionItem {
   parentAssemblyBinding: string
 }
 
+const isAssemblyOptionToggle = compose<AssemblyOption, string, boolean>(
+  equals(CHOICE_TYPES.TOGGLE),
+  getItemChoiceType
+)
+
 export const buildRemovedOptions = (
   item: OrderFormItem,
-  orderForm: any,
+  orderForm: OrderForm,
   assemblyOptionsMap: Record<string, AssemblyOption[]>
 ): RemovedItem[] => {
   const assemblyOptions = assemblyOptionsMap[item.id]
   if (!assemblyOptions) {
     return []
   }
+  // For now, it makes sense it should only work for toggle type of assembly options
+  const onlyToggleAssemblies = assemblyOptions.filter(isAssemblyOptionToggle)
+
   const itemsWithInitials: InitialItem[] = []
-  for (const assemblyOption of assemblyOptions) {
+  for (const assemblyOption of onlyToggleAssemblies) {
     if (assemblyOption.composition) {
       for (const compItem of assemblyOption.composition.items) {
         if (compItem.initialQuantity > 0) {