+++ node/typings/Catalog.ts
@@ -60,6 +60,11 @@ interface Product {
   link: string
   description: string
   items: Item[]
+  itemMetadata: {
+    items: CatalogMetadataItem[]
+  }
+  titleTag: string
+  jsonSpecifications: string
 }
 
 interface Item {
@@ -78,6 +83,9 @@ interface Item {
     imageTag: string
     imageUrl: string
     imageText: string
+  }[]
+  videos: {
+    videoUrl: string
   }[]
   variations: string[]
   sellers: Seller[]
@@ -103,7 +111,7 @@ interface Seller {
       Name: string
     }[]
     DiscountHighLight: any[]
-    GiftSkuIds: any[]
+    GiftSkuIds: string[]
     Teasers: any[]
     BuyTogether: any[]
     ItemMetadataAttachment: any[]